task6 done
