+++ lab3.py
@@ -1,4 +1,5 @@
 import math
+from square_generator import SquareGenerator
 
 # task1
 numlist = [x ** 2 for x in range(11)]
@@ -13,15 +14,6 @@ def e_squares(start, end):
 print(e_squares(2, 10))
 
 
-# task3
-
-class SquareGenerator:
-    def squares(self, start, end):
-        if end < start:
-            raise ValueError("End of the range cannot be less than the start.")
-
-        return [x ** 2 for x in range(start, end + 1)]
-
 
 a = SquareGenerator()
 print(a.squares(3, 10))
@@ -35,4 +27,6 @@ print(rootlist)
 
 # task5
 a.squares(10, 1)
+
+
